Add support in suffix / prefix per processed file
Currently suffix / prefix options are passed as strings and are being applied to every processed file. This change allows passing an these options as an object, where the key is the file to which this option applies to and the value is the prefix/suffix to add. e.g:

suffix: {
  '../path/to/file1': 'obj1 = getMessages();',
  '../path/to/file2': 'obj2 = getMessages();'
}